ai/requirements.txt
langchain-core
langchain-groq
langgraph
fastapi
uvicorn[standard]
python-dotenv
pydantic
pinecone 
huggingface-hub
motor
pymongo